feat(transaction): Membuat transaksi

+++ src/controllers/event/event.controller.ts
@@ -134,6 +134,10 @@ export const updateEvent = async (req: Request, res: Response) => {
     const { id } = req.params;
     const data = req.body;
 
+    const event = await prisma.event.findUnique({ where: { id: Number(id) } });
+
+    if (!event) return res.status(404).json({ error: "Event not found" });
+
     const updatedEvent = await prisma.event.update({
       where: { id: Number(id) },
       data,
@@ -150,6 +154,8 @@ export const updateEvent = async (req: Request, res: Response) => {
 export const deleteEvent = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const event = await prisma.event.findUnique({ where: { id: Number(id) } });
+    if (!event) return res.status(404).json({ error: "Event not found" });
 
     await prisma.event.delete({ where: { id: Number(id) } });
